+++ src/app/SettingsUseCaseFactory.java
@@ -1,6 +1,5 @@
 package app;
 
-import interface_adapter.ViewManagerModel;
 import interface_adapter.logged_in.LoggedInViewModel;
 import interface_adapter.settings.SettingsController;
 import interface_adapter.settings.SettingsPresenter;
@@ -48,7 +47,8 @@ public class SettingsUseCaseFactory {
       LoginUserDataAccessInterface userDao)
       throws IOException {
 
-    SettingsOutputBoundary settingsOutputBoundary = new SettingsPresenter(settingsViewModel, loggedInViewModel);
+    SettingsOutputBoundary settingsOutputBoundary =
+        new SettingsPresenter(settingsViewModel, loggedInViewModel);
     SettingsInputBoundary settingsInteractor =
         new SettingsInteractor(
             userSettingsDataAccessObject,
