+++ src/use_case/settings/SettingsOutputData.java
@@ -1,12 +1,15 @@
 package use_case.settings;
 
 public class SettingsOutputData {
-  private String error;
+  private final String error;
 
   public SettingsOutputData(String error) {
     this.error = error;
   }
 
+  public boolean getIsError() {
+    return error != null;
+  }
 
   public String getError() {
     return error;
