+++ tests/view/RoomsTest.java
@@ -193,7 +193,8 @@ public class RoomsTest extends ButtonTest {
             new FirebaseUserDataAccessObject(client),
             roomsViewModel,
             null,
-            startSearchController, null);
+            startSearchController,
+            null);
     RoomsState currentstate = roomsViewModel.getState();
     currentstate.setUser(createDummyUser());
     currentstate.setRoomUid(createDummyRoom().getUid());
